backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-docx>=1.1.0
openai>=1.40.0
anthropic>=0.30.0
python-dotenv>=1.0.0
